Avoid using anonymous classes for quick-fixes
An anonymous class creates a risk of accidentally capturing a variable
 which is bad for a quick fix
