+++ src/main/kotlin/org/rust/ide/fixes/ReloadProjectQuickFix.kt
@@ -1,0 +1,23 @@
+/*
+ * Use of this source code is governed by the MIT license that can be
+ * found in the LICENSE file.
+ */
+
+package org.rust.ide.fixes
+
+import com.intellij.codeInspection.LocalQuickFix
+import com.intellij.codeInspection.ProblemDescriptor
+import com.intellij.openapi.project.Project
+import org.rust.cargo.project.model.cargoProjects
+import org.rust.ide.notifications.confirmLoadingUntrustedProject
+import org.rust.openapiext.saveAllDocuments
+
+class ReloadProjectQuickFix : LocalQuickFix {
+    override fun getFamilyName(): String = "Reload project"
+    override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
+        // It seems it's impossible to be here if project is not trusted but let's be 100% sure
+        if (!project.confirmLoadingUntrustedProject()) return
+        saveAllDocuments()
+        project.cargoProjects.refreshAllProjects()
+    }
+}
