+++ src/main/kotlin/org/rust/ide/inspections/RsExtraSemicolonInspection.kt
@@ -5,11 +5,8 @@
 
 package org.rust.ide.inspections
 
-import com.intellij.codeInspection.LocalQuickFix
-import com.intellij.codeInspection.ProblemDescriptor
-import com.intellij.openapi.project.Project
+import org.rust.ide.fixes.RemoveSemicolonFix
 import org.rust.lang.core.dfa.ExitPoint
-import org.rust.lang.core.psi.RsExprStmt
 import org.rust.lang.core.psi.RsFunction
 import org.rust.lang.core.psi.RsVisitor
 import org.rust.lang.core.types.normType
@@ -27,30 +24,19 @@ class RsExtraSemicolonInspection : RsLocalInspectionTool() {
 
     override fun buildVisitor(holder: RsProblemsHolder, isOnTheFly: Boolean): RsVisitor =
         object : RsWithMacrosInspectionVisitor() {
-            override fun visitFunction2(o: RsFunction) = inspect(holder, o)
-        }
-}
-
-
-private fun inspect(holder: RsProblemsHolder, fn: RsFunction) {
-    val retType = fn.retType?.typeReference ?: return
-    if (retType.normType is TyUnit) return
-    ExitPoint.process(fn) { exitPoint ->
-        if (exitPoint is ExitPoint.InvalidTailStatement) {
-            holder.registerProblem(
-                exitPoint.stmt,
-                "Function returns () instead of ${retType.text}",
-                object : LocalQuickFix {
-                    override fun getName() = "Remove semicolon"
-
-                    override fun getFamilyName() = name
-
-                    override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
-                        val statement = (descriptor.psiElement as RsExprStmt)
-                        statement.semicolon?.delete()
+            override fun visitFunction2(o: RsFunction) {
+                val retType = o.retType?.typeReference ?: return
+                if (retType.normType is TyUnit) return
+                ExitPoint.process(o) { exitPoint ->
+                    if (exitPoint is ExitPoint.InvalidTailStatement) {
+                        holder.registerProblem(
+                            exitPoint.stmt,
+                            "Function returns () instead of ${retType.text}",
+                            RemoveSemicolonFix()
+                        )
                     }
                 }
-            )
+            }
         }
-    }
 }
+
