+++ src/main/kotlin/org/rust/ide/inspections/RsFieldInitShorthandInspection.kt
@@ -5,14 +5,11 @@
 
 package org.rust.ide.inspections
 
-import com.intellij.codeInspection.LocalQuickFix
-import com.intellij.codeInspection.ProblemDescriptor
 import com.intellij.codeInspection.ProblemHighlightType
-import com.intellij.openapi.project.Project
+import org.rust.ide.fixes.ChangeToFieldShorthandFix
 import org.rust.lang.core.psi.RsPathExpr
 import org.rust.lang.core.psi.RsStructLiteralField
 import org.rust.lang.core.psi.RsVisitor
-
 
 class RsFieldInitShorthandInspection : RsLocalInspectionTool() {
     override fun buildVisitor(holder: RsProblemsHolder, isOnTheFly: Boolean): RsVisitor = object : RsWithMacrosInspectionVisitor() {
@@ -23,23 +20,10 @@ class RsFieldInitShorthandInspection : RsLocalInspectionTool() {
                 o,
                 "Expression can be simplified",
                 ProblemHighlightType.WEAK_WARNING,
-                object : LocalQuickFix {
-                    override fun getFamilyName(): String = "Use initialization shorthand"
-
-                    override fun applyFix(project: Project, descriptor: ProblemDescriptor) {
-                        applyShorthandInit(descriptor.psiElement as RsStructLiteralField)
-                    }
-                }
+                ChangeToFieldShorthandFix()
             )
         }
     }
 
     override val isSyntaxOnly: Boolean = true
-
-    companion object {
-        fun applyShorthandInit(field: RsStructLiteralField) {
-            field.expr?.delete()
-            field.colon?.delete()
-        }
-    }
 }
