+++ src/main/kotlin/org/rust/ide/refactoring/inlineFunction/RsInlineFunctionProcessor.kt
@@ -18,9 +18,9 @@ import com.intellij.refactoring.BaseRefactoringProcessor
 import com.intellij.usageView.UsageInfo
 import com.intellij.usageView.UsageViewDescriptor
 import com.intellij.util.containers.MultiMap
+import org.rust.ide.fixes.ChangeToFieldShorthandFix
 import org.rust.ide.fixes.deleteUseSpeck
 import org.rust.ide.fixes.updateMutable
-import org.rust.ide.inspections.RsFieldInitShorthandInspection
 import org.rust.ide.refactoring.RsInlineUsageViewDescriptor
 import org.rust.ide.refactoring.freshenName
 import org.rust.ide.refactoring.inlineTypeAlias.fillPathWithActualType
@@ -472,7 +472,7 @@ private class InlineSingleUsage(
         for (field in function.descendantsOfType<RsStructLiteralField>()) {
             val name = field.identifier?.text ?: continue
             if (field.colon != null && name == field.expr?.text) {
-                RsFieldInitShorthandInspection.applyShorthandInit(field)
+                ChangeToFieldShorthandFix.applyShorthandInit(field)
             }
         }
     }
